+++ app/helpers/assigned_courses_helper.rb
@@ -1,0 +1,2 @@
+module AssignedCoursesHelper
+end