+++ db/migrate/20240220043107_create_progresses.rb
@@ -1,0 +1,9 @@
+class CreateProgresses < ActiveRecord::Migration[7.1]
+  def change
+    create_table :progresses do |t|
+      t.string :name
+
+      t.timestamps
+    end
+  end
+end