+++ db/migrate/20240220052052_add_row_order_to_progress_and_assigned_course.rb
@@ -1,0 +1,6 @@
+class AddRowOrderToProgressAndAssignedCourse < ActiveRecord::Migration[7.1]
+  def change
+    add_column :progresses, :row_order, :integer
+    add_column :assigned_courses, :row_order, :integer
+  end
+end