+++ db/migrate/20240220070657_rename_progress_column_in_assigned_courses.rb
@@ -1,0 +1,5 @@
+class RenameProgressColumnInAssignedCourses < ActiveRecord::Migration[7.1]
+  def change
+    rename_column :assigned_courses, :progress, :progress_status
+  end
+end