+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.1].define(version: 2024_02_19_175618) do
+ActiveRecord::Schema[7.1].define(version: 2024_02_20_070657) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
@@ -52,6 +52,22 @@ ActiveRecord::Schema[7.1].define(version: 2024_02_19_175618) do
     t.index ["blob_id", "variation_digest"], name: "index_active_storage_variant_records_uniqueness", unique: true
   end
 
+  create_table "assigned_courses", force: :cascade do |t|
+    t.boolean "pass", default: false, null: false
+    t.float "assessment_score"
+    t.datetime "date_completed"
+    t.integer "progress_status", default: 0
+    t.bigint "user_id", null: false
+    t.bigint "course_id", null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.bigint "progress_id"
+    t.integer "row_order"
+    t.index ["course_id"], name: "index_assigned_courses_on_course_id"
+    t.index ["progress_id"], name: "index_assigned_courses_on_progress_id"
+    t.index ["user_id"], name: "index_assigned_courses_on_user_id"
+  end
+
   create_table "courses", force: :cascade do |t|
     t.string "title"
     t.bigint "user_id", null: false
@@ -64,6 +80,13 @@ ActiveRecord::Schema[7.1].define(version: 2024_02_19_175618) do
     t.float "impact", default: 0.0
     t.index ["slug"], name: "index_courses_on_slug", unique: true
     t.index ["user_id"], name: "index_courses_on_user_id"
+  end
+
+  create_table "progresses", force: :cascade do |t|
+    t.string "name"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.integer "row_order"
   end
 
   create_table "users", force: :cascade do |t|
@@ -87,5 +110,8 @@ ActiveRecord::Schema[7.1].define(version: 2024_02_19_175618) do
 
   add_foreign_key "active_storage_attachments", "active_storage_blobs", column: "blob_id"
   add_foreign_key "active_storage_variant_records", "active_storage_blobs", column: "blob_id"
+  add_foreign_key "assigned_courses", "courses"
+  add_foreign_key "assigned_courses", "progresses"
+  add_foreign_key "assigned_courses", "users"
   add_foreign_key "courses", "users"
 end