+++ test/controllers/assigned_courses_controller_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class AssignedCoursesControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end